+++ packages/headless/src/controllers/breadcrumb-manager/headless-breadcrumb-manager.test.ts
@@ -93,10 +93,10 @@ describe('headless breadcrumb manager', () => {
     state.search.response.facets = [
       buildMockNumericFacetResponse({facetId, values: [mockValue]}),
     ];
-    expect(breadcrumbManager.hasBreadcrumbs()).toBe(true);
+    expect(breadcrumbManager.state.hasBreadcrumbs).toBe(true);
   });
 
   it('hasBreadcrumbs returns false when no facet value is selected', () => {
-    expect(breadcrumbManager.hasBreadcrumbs()).toBe(false);
+    expect(breadcrumbManager.state.hasBreadcrumbs).toBe(false);
   });
 });