+++ packages/headless/src/features/did-you-mean/did-you-mean-state.ts
@@ -1,9 +1,23 @@
 import {QueryCorrection} from '../../api/search/search/query-corrections';
 
 export interface DidYouMeanState {
+  /**
+   * Specifies if the "did you mean" feature of the Coveo platform should be enabled or not.
+   */
   enableDidYouMean: boolean;
+  /**
+   * The correction that was applied to the query. If no correction was applied, will default to an empty string.
+   */
   wasCorrectedTo: string;
+  /**
+   * Specifies if the query was automatically corrected by Headless.
+   *
+   * This happens when there is no result returned by the API for a particular mispelling.
+   */
   wasAutomaticallyCorrected: boolean;
+  /**
+   * The query correction that is currently applied by the "did you mean" module.
+   */
   queryCorrection: QueryCorrection;
 }