+++ packages/headless/src/features/search-parameters/search-parameter-serializer.test.ts
@@ -1,3 +1,5 @@
+import {buildDateRange} from '../../controllers/facets/range-facet/date-facet/headless-date-facet';
+import {buildNumericRange} from '../../controllers/facets/range-facet/numeric-facet/headless-numeric-facet';
 import {buildMockSearchParameters} from '../../test/mock-search-parameters';
 import {buildSearchParameterSerializer} from './search-parameter-serializer';
 
@@ -32,6 +34,71 @@ describe('buildSearchParameterSerializer', () => {
       const f = {author: [1]} as never;
       const result = serialize({f});
       expect(result).toEqual('');
+    });
+
+    it('serializes the #nf parameter correctly', () => {
+      const nf = {
+        size: [
+          buildNumericRange({start: 0, end: 10}),
+          buildNumericRange({start: 10, end: 20}),
+        ],
+        amount: [buildNumericRange({start: 100, end: 200})],
+      };
+
+      const result = serialize({nf});
+      expect(result).toEqual('nf[size]=0..10,10..20&nf[amount]=100..200');
+    });
+
+    describe('when the #nf parameter contains facetIds with invalid values', () => {
+      it('if a range does not have a #start key, it does not include it', () => {
+        const nf = {size: [{}]} as never;
+        const result = serialize({nf});
+        expect(result).toEqual('');
+      });
+
+      it('if a range does not have an #end key, it does not include it', () => {
+        const nf = {size: [{start: 10}]} as never;
+        const result = serialize({nf});
+        expect(result).toEqual('');
+      });
+    });
+
+    it('serializes the #df parameter correctly', () => {
+      const date1 = '2010/01/01@05:00:00';
+      const date2 = '2011/01/01@05:00:00';
+      const date3 = '2012/01/01@05:00:00';
+
+      const df = {
+        date: [
+          buildDateRange({start: date1, end: date2, useLocalTime: true}),
+          buildDateRange({start: date2, end: date3, useLocalTime: true}),
+        ],
+        created: [
+          buildDateRange({start: date1, end: date2, useLocalTime: true}),
+        ],
+      };
+
+      const result = serialize({df});
+      const range1 = `${date1}..${date2}`;
+      const range2 = `${date2}..${date3}`;
+
+      expect(result).toEqual(
+        `df[date]=${range1},${range2}&df[created]=${range1}`
+      );
+    });
+
+    describe('when the #df parameter contains facetIds with invalid values', () => {
+      it('if a range does not have a #start key, it does not include it', () => {
+        const df = {created: [{}]} as never;
+        const result = serialize({df});
+        expect(result).toEqual('');
+      });
+
+      it('if a range does not have an #end key, it does not include it', () => {
+        const df = {size: [{start: '2010/01/01@05:00:00'}]} as never;
+        const result = serialize({df});
+        expect(result).toEqual('');
+      });
     });
   });
 
@@ -78,12 +145,153 @@ describe('buildSearchParameterSerializer', () => {
         },
       });
     });
+
+    it('deserializes two category facets correctly', () => {
+      const result = deserialize('cf[author]=a,b&cf[filetype]=c,d');
+      expect(result).toEqual({
+        cf: {
+          author: ['a', 'b'],
+          filetype: ['c', 'd'],
+        },
+      });
+    });
+
+    it('deserializes a numeric facet with multiple selections', () => {
+      const result = deserialize('nf[size]=0..10,10..20');
+      expect(result).toEqual({
+        nf: {
+          size: [
+            buildNumericRange({start: 0, end: 10, state: 'selected'}),
+            buildNumericRange({start: 10, end: 20, state: 'selected'}),
+          ],
+        },
+      });
+    });
+
+    it('deserializes multiple numeric facets with selected values', () => {
+      const result = deserialize('nf[size]=0..10&nf[amount]=100..200');
+      expect(result).toEqual({
+        nf: {
+          size: [buildNumericRange({start: 0, end: 10, state: 'selected'})],
+          amount: [
+            buildNumericRange({start: 100, end: 200, state: 'selected'}),
+          ],
+        },
+      });
+    });
+
+    it('deserializes a numeric facet with a float range', () => {
+      const result = deserialize('nf[size]=7.5..8.5,8.5..9.5');
+      expect(result).toEqual({
+        nf: {
+          size: [
+            buildNumericRange({start: 7.5, end: 8.5, state: 'selected'}),
+            buildNumericRange({start: 8.5, end: 9.5, state: 'selected'}),
+          ],
+        },
+      });
+    });
+
+    it(`when a numeric facet range contains a non-numeric value,
+    the deserializer sets the values for the facetId to an empty array`, () => {
+      const result = deserialize('nf[size]=0..a');
+      expect(result).toEqual({
+        nf: {size: []},
+      });
+    });
+
+    it(`when a numeric facet range contains one number,
+    the deserializer sets the values for the facetId to an empty array`, () => {
+      const result = deserialize('nf[size]=0');
+      expect(result).toEqual({
+        nf: {size: []},
+      });
+    });
+
+    it('deserializes a date facet with multiple selections', () => {
+      const date1 = '2010/01/01@05:00:00';
+      const date2 = '2011/01/01@05:00:00';
+      const date3 = '2012/01/01@05:00:00';
+
+      const rangeA = `${date1}..${date2}`;
+      const rangeB = `${date2}..${date3}`;
+
+      const result = deserialize(`df[date]=${rangeA},${rangeB}`);
+
+      expect(result).toEqual({
+        df: {
+          date: [
+            buildDateRange({
+              start: date1,
+              end: date2,
+              useLocalTime: true,
+              state: 'selected',
+            }),
+            buildDateRange({
+              start: date2,
+              end: date3,
+              useLocalTime: true,
+              state: 'selected',
+            }),
+          ],
+        },
+      });
+    });
+
+    it('deserializes multiple date facets with selected values', () => {
+      const date1 = '2010/01/01@05:00:00';
+      const date2 = '2011/01/01@05:00:00';
+      const range = `${date1}..${date2}`;
+
+      const result = deserialize(`df[date]=${range}&df[created]=${range}`);
+      const expected = buildDateRange({
+        start: date1,
+        end: date2,
+        useLocalTime: true,
+        state: 'selected',
+      });
+
+      expect(result).toEqual({
+        df: {
+          date: [expected],
+          created: [expected],
+        },
+      });
+    });
+
+    it(`when a date facet range contains a invalid format,
+    the deserializer sets the values for the facetId to an empty array`, () => {
+      const result = deserialize('df[date]=2010/01/01@05:00:00..a');
+      expect(result).toEqual({
+        df: {date: []},
+      });
+    });
+
+    it(`when a numeric facet range contains one number,
+    the deserializer sets the values for the facetId to an empty array`, () => {
+      const result = deserialize('df[date]=2010/01/01@05:00:00');
+      expect(result).toEqual({
+        df: {date: []},
+      });
+    });
   });
 
   it('can serialize and deserialize all search parameters', () => {
     const f = {author: ['a', 'b']};
     const cf = {geography: ['a', 'b']};
-    const parameters = buildMockSearchParameters({f, cf});
+    const nf = {
+      size: [buildNumericRange({start: 0, end: 10, state: 'selected'})],
+    };
+    const df = {
+      created: [
+        buildDateRange({
+          start: '2010/01/01@05:00:00',
+          end: '2011/01/01@05:00:00',
+          state: 'selected',
+        }),
+      ],
+    };
+    const parameters = buildMockSearchParameters({f, cf, nf, df});
 
     const {serialize, deserialize} = buildSearchParameterSerializer();
     const serialized = serialize(parameters);